Add new 'Login' view
Create 'login.js' in which users are able to submit a username and 
password, which is sent to the backend for validation. (The backend 
returns a token for the user, which is not currently implemented.) 
Update 'wall.js' and 'App.js' to provide a route and anchor to the new 
'Login' view.